+++ index.php
@@ -40,25 +40,9 @@ $cards = [
         'avatar' => 'userpic.jpg'
     ],
 ];
-function truncate_text(string $text, int $truncate_length = 300)
-{
-    if (mb_strlen($text) <= $truncate_length) {
-        $final_text = $text;
-    } else {
-        $words = explode(" ", $text);
-        $i = 0;
-        while ($current_length - 1 < $truncate_length) {
-            $current_length += mb_strlen($words[$i]);
-            $current_length++;
-            $i++;
-        }
-        $final_text = implode(" ", array_slice($words, 0, $i - 1));
-        $final_text .= '...';
-    }
-    return $final_text;
-}
 
 require_once('helpers.php');
+require_once('functions.php');
 $page_content = include_template('main.php', ['cards' => $cards]);
 $layout_content = include_template('layout.php', ['content' => $page_content, 'user' => $user_name, 'title' => $title, 'is_auth' => $is_auth]);
 print($layout_content);
